% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{create_service}
\alias{create_service}
\title{Creates a service, which defines the configuration for the following entities:}
\usage{
create_service(Name, CreatorRequestId = NULL, Description = NULL,
  DnsConfig, HealthCheckConfig = NULL, HealthCheckCustomConfig = NULL)
}
\arguments{
\item{Name}{[required] The name that you want to assign to the service.}

\item{CreatorRequestId}{A unique string that identifies the request and that allows failed \code{CreateService} requests to be retried without the risk of executing the operation twice. \code{CreatorRequestId} can be any unique string, for example, a date/time stamp.}

\item{Description}{A description for the service.}

\item{DnsConfig}{[required] A complex type that contains information about the records that you want Route 53 to create when you register an instance.}

\item{HealthCheckConfig}{\emph{Public DNS namespaces only.} A complex type that contains settings for an optional health check. If you specify settings for a health check, Route 53 associates the health check with all the records that you specify in \code{DnsConfig}.

For information about the charges for health checks, see \href{http://aws.amazon.com/route53/pricing}{Route 53 Pricing}.}

\item{HealthCheckCustomConfig}{}
}
\description{
Creates a service, which defines the configuration for the following entities:
}
\details{
\itemize{
\item Up to three records (A, AAAA, and SRV) or one CNAME record
\item Optionally, a health check
}

After you create the service, you can submit a RegisterInstance request, and Amazon Route 53 uses the values in the configuration to create the specified entities.

For the current limit on the number of instances that you can register using the same namespace and using the same service, see \href{http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/DNSLimitations.html#limits-api-entities-autonaming}{Limits on Auto Naming} in the \emph{Route 53 Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{create_service(
  Name = "string",
  CreatorRequestId = "string",
  Description = "string",
  DnsConfig = list(
    NamespaceId = "string",
    RoutingPolicy = "MULTIVALUE"|"WEIGHTED",
    DnsRecords = list(
      list(
        Type = "SRV"|"A"|"AAAA"|"CNAME",
        TTL = 123
      )
    )
  ),
  HealthCheckConfig = list(
    Type = "HTTP"|"HTTPS"|"TCP",
    ResourcePath = "string",
    FailureThreshold = 123
  ),
  HealthCheckCustomConfig = list(
    FailureThreshold = 123
  )
)
}
}

