% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{delete_namespace}
\alias{delete_namespace}
\title{Deletes a namespace from the current account}
\usage{
delete_namespace(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] The ID of the namespace that you want to delete.}
}
\description{
Deletes a namespace from the current account. If the namespace still contains one or more services, the request fails.
}
\examples{

}
