% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{delete_service}
\alias{delete_service}
\title{Deletes a specified service}
\usage{
delete_service(Id)
}
\arguments{
\item{Id}{The ID of the service that you want to delete.}
}
\description{
Deletes a specified service. If the service still contains one or more registered instances, the request fails.
}
\examples{

}
