% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{deregister_instance}
\alias{deregister_instance}
\title{Deletes the Amazon Route 53 DNS records and health check, if any, that AWS Cloud Map created for the specified instance}
\usage{
deregister_instance(
  ServiceId = "string",
  InstanceId = "string"
)
}
\arguments{
\item{ServiceId}{[required] The ID of the service that the instance is associated with.}

\item{InstanceId}{[required] The value that you specified for \code{Id} in the RegisterInstance request.}
}
\description{
Deletes the Amazon Route 53 DNS records and health check, if any, that AWS Cloud Map created for the specified instance.
}
\examples{

}
