% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{get_instance}
\alias{get_instance}
\title{Gets information about a specified instance}
\usage{
get_instance(
  ServiceId = "string",
  InstanceId = "string"
)
}
\arguments{
\item{ServiceId}{[required] The ID of the service that the instance is associated with.}

\item{InstanceId}{[required] The ID of the instance that you want to get information about.}
}
\description{
Gets information about a specified instance.
}
\examples{

}
