% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{get_instances_health_status}
\alias{get_instances_health_status}
\title{Gets the current health status (\code{Healthy}, \code{Unhealthy}, or \code{Unknown}) of one or more instances that are associated with a specified service}
\usage{
get_instances_health_status(ServiceId, Instances = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ServiceId}{The ID of the service that the instance is associated with.}

\item{Instances}{An array that contains the IDs of all the instances that you want to get the health status for.

If you omit \code{Instances}, Amazon Route 53 returns the health status for all the instances that are associated with the specified service.

To get the IDs for the instances that you've registered by using a specified service, submit a ListInstances request.}

\item{MaxResults}{The maximum number of instances that you want Route 53 to return in the response to a \code{GetInstancesHealthStatus} request. If you don't specify a value for \code{MaxResults}, Route 53 returns up to 100 instances.}

\item{NextToken}{For the first \code{GetInstancesHealthStatus} request, omit this value.

If more than \code{MaxResults} instances match the specified criteria, you can submit another \code{GetInstancesHealthStatus} request to get the next group of results. Specify the value of \code{NextToken} from the previous response in the next request.}
}
\description{
Gets the current health status (\code{Healthy}, \code{Unhealthy}, or \code{Unknown}) of one or more instances that are associated with a specified service.
}
\details{
There is a brief delay between when you register an instance and when the health status for the instance is available.
}
\examples{

}
