% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{get_namespace}
\alias{get_namespace}
\title{Gets information about a namespace}
\usage{
get_namespace(Id)
}
\arguments{
\item{Id}{The ID of the namespace that you want to get information about.}
}
\description{
Gets information about a namespace.
}
\examples{

}
