% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{get_operation}
\alias{get_operation}
\title{Gets information about any operation that returns an operation ID in the response, such as a CreateService request}
\usage{
get_operation(OperationId)
}
\arguments{
\item{OperationId}{[required] The ID of the operation that you want to get more information about.}
}
\description{
Gets information about any operation that returns an operation ID in the response, such as a \code{CreateService} request.
}
\details{
To get a list of operations that match specified criteria, see ListOperations.
}
\section{Accepted Parameters}{
\preformatted{get_operation(
  OperationId = "string"
)
}
}

