% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{get_service}
\alias{get_service}
\title{Gets the settings for a specified service}
\usage{
get_service(Id)
}
\arguments{
\item{Id}{[required] The ID of the service that you want to get settings for.}
}
\description{
Gets the settings for a specified service.
}
\section{Accepted Parameters}{
\preformatted{get_service(
  Id = "string"
)
}
}

