% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_package.R
\docType{package}
\name{paws.servicediscovery-package}
\alias{paws.servicediscovery}
\alias{paws.servicediscovery-package}
\title{paws.servicediscovery: AWS Cloud Map}
\description{
AWS Cloud Map lets you configure public DNS,
    private DNS, or HTTP namespaces that your microservice applications
    run in. When an instance of the service becomes available, you can
    call the AWS Cloud Map API to register the instance with AWS Cloud
    Map. For public or private DNS namespaces, AWS Cloud Map automatically
    creates DNS records and an optional health check. Clients that submit
    public or private DNS queries, or HTTP requests, for the service
    receive an answer that contains up to eight healthy records.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
