% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_operations.R
\name{deregister_instance}
\alias{deregister_instance}
\title{Deletes the records and the health check, if any, that Amazon Route 53 created for the specified instance}
\usage{
deregister_instance(ServiceId, InstanceId)
}
\arguments{
\item{ServiceId}{The ID of the service that the instance is associated with.}

\item{InstanceId}{The value that you specified for \code{Id} in the RegisterInstance request.}
}
\description{
Deletes the records and the health check, if any, that Amazon Route 53 created for the specified instance.
}
\examples{

}
