% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.servicediscovery_package.R
\docType{package}
\name{paws.servicediscovery-package}
\alias{paws.servicediscovery}
\alias{paws.servicediscovery-package}
\title{paws.servicediscovery: Amazon Route 53 Auto Naming}
\description{
Amazon Route 53 auto naming lets you configure
public or private namespaces that your microservice applications run
in. When instances of the service become available, you can call the
auto naming API to register the instance, and Route 53 automatically
creates up to five DNS records and an optional health check. Clients
that submit DNS queries for the service receive an answer that
contains up to eight healthy records.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
